Ver 1.9.4

+++ verify/AOJ-GRL-1-C.test.cpp
@@ -9,14 +9,14 @@ using namespace std;
 int main(){
     int V, E;
     cin >> V >> E;
-    Graph<int> G(V, true);
+    Graph<long long> G(V, true);
     for(int i = 0; i < E; ++i){
         int s, t, d;
         cin >> s >> t >> d;
         G.add(s, t, d);
     }
 
-    WarshallFloyd<int> wf(G);
+    WarshallFloyd<long long> wf(G);
     if(wf.negative){
         cout << "NEGATIVE CYCLE\n";
     }
